refactor(constants): unify MAX_FLOOR on u32; drop world.rs duplicate
`MAX_FLOOR` lived in two places: `constants.rs` as `i32: 5` and
`world.rs` as `u32: 5`. Same value, different types, drift hazard.

Unifies on `u32` in `constants.rs` since floor indices are inherently
non-negative and `Floor(u32)` is the canonical floor type throughout
the game crate. `map.depth: i32` is an engine quirk (bracket-lib
convention); the handful of sites that compare against it now cast
explicitly:

- `dungeon.rs::snapshot_floor` warning gate: `(map.depth as u32) < MAX_FLOOR`
- `exit_points::DistantExit::build_map`: `build_data.map.depth as u32 >= MAX_FLOOR`
- `forest::deepest_forest_depth()` returns `(MAX_FLOOR - 1) as i32`
- `forest.rs` test helpers: `(MAX_FLOOR - 1) as i32`
- `temple.rs` test helper: `MAX_FLOOR as i32`
- `assets/mod.rs` spawn-table test: `2..MAX_FLOOR as i32` to match
  `MonsterSpawnInfo.{min,max}_floor: i32`

Cleanups in the other direction:
- `tile.rs::chasm_fall_reaction_system`: drops the now-redundant
  `as u32` cast — `MAX_FLOOR` is u32 directly.
- `world.rs`: deletes its own `pub const MAX_FLOOR: u32 = 5;` +
  doc comment, imports the canonical one from `constants.rs`.

Doc comment on `constants::MAX_FLOOR` explains the type choice + the
`map.depth` cast pattern so future readers know why the casts are
there.

717 tests pass.

+++ src/assets/mod.rs
@@ -1267,7 +1267,7 @@ mod species_tests {
 
         // Forest floors are 2..=MAX_FLOOR-1. Floor 1 is the town (no
         // spawns by design); MAX_FLOOR is the temple (no spawns yet).
-        for floor in 2..crate::constants::MAX_FLOOR {
+        for floor in 2..crate::constants::MAX_FLOOR as i32 {
             let count = table
                 .spawns
                 .iter()
